Allow sharing of overview components
