--- app/ui/browser-blueprint/views/overview/cards/overview-card.jsx
@@ -1,42 +1,0 @@
-/*
-Copyright 2016 Mozilla
-
-Licensed under the Apache License, Version 2.0 (the "License"); you may not use
-this file except in compliance with the License. You may obtain a copy of the
-License at http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software distributed
-under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-CONDITIONS OF ANY KIND, either express or implied. See the License for the
-specific language governing permissions and limitations under the License.
-*/
-
-import React, { PropTypes } from 'react';
-
-/**
- * A factory class used to instantiate Cards displayed in the overview.
- * Calls the appropriate underlying card subclass, depending on metadata.
- */
-import * as SharedPropTypes from '../../../model/shared-prop-types';
-import SimpleCard from './simple-card';
-import ProductCard from './product-card';
-
-const OverviewCard = function(props) {
-  const meta = props.page.meta;
-
-  if (meta.rating || meta.price) {
-    return (<ProductCard {...props} />);
-  }
-
-  return (<SimpleCard {...props} />);
-};
-
-OverviewCard.displayName = 'OverviewCard';
-
-OverviewCard.propTypes = {
-  page: SharedPropTypes.Page.isRequired,
-  pageIndex: PropTypes.number.isRequired,
-  isSelected: PropTypes.bool.isRequired,
-  onClick: PropTypes.func.isRequired,
-};
-
-export default OverviewCard;
