+++ app/ui/shared/widgets/overview/summaries/product-summary.jsx
@@ -13,10 +13,12 @@ specific language governing permissions and limitations under the License.
 import React, { PropTypes } from 'react';
 
 import Style from '../../../../shared/style';
+import Ratings from './ratings';
 
-const GRADIENT_MASK = `-webkit-gradient(linear, center top, center bottom,
-                       color-stop(0.8, rgba(255,255,255,1)),
-                       color-stop(1, rgba(255,255,255,0)))`;
+const GRADIENT_MASK = `
+  -webkit-gradient(linear, center top, center bottom,
+  color-stop(0.8, rgba(255,255,255,1)),
+  color-stop(1, rgba(255,255,255,0)))`;
 
 const CONTAINER_STYLE = Style.registerStyle({
   flex: 1,
@@ -30,10 +32,19 @@ const TITLE_STYLE = Style.registerStyle({
   textOverflow: 'ellipsis',
   paddingBottom: '1px',
   color: 'var(--theme-overview-summary-title-color)',
-  '-webkit-mask-image': GRADIENT_MASK,
+  WebkitMaskImage: GRADIENT_MASK,
 });
 
-const URL_STYLE = Style.registerStyle({
+const PRICE_STYLE = Style.registerStyle({
+  fontWeight: 'bold',
+  flexShrink: 0,
+  display: 'block',
+  overflow: 'hidden',
+  textOverflow: 'ellipsis',
+  whiteSpace: 'nowrap',
+});
+
+const REVIEW_COUNT_STYLE = Style.registerStyle({
   flexShrink: 0,
   display: 'block',
   overflow: 'hidden',
@@ -43,24 +54,46 @@ const URL_STYLE = Style.registerStyle({
   color: 'var(--theme-overview-summary-subtitle-color)',
 });
 
-const SimpleSummary = function(props) {
+const ProductSummary = props => {
+  const priceEl = props.price ? (
+    <div className={PRICE_STYLE}>
+      {props.price}
+    </div>
+  ) : undefined;
+
+  const ratingsEl = props.rating ? (
+    <Ratings rating={props.rating}
+      maxRating={props.maxRating}
+      minRating={props.minRating} />
+  ) : undefined;
+
+  const reviewEl = props.reviewCount ? (
+    <div className={REVIEW_COUNT_STYLE}>
+      {`${props.reviewCount} reviews`}
+    </div>
+  ) : undefined;
+
   return (
     <div className={CONTAINER_STYLE}>
       <div className={TITLE_STYLE}>
         {props.title}
       </div>
-      <div className={URL_STYLE}>
-        {props.url}
-      </div>
+      {priceEl}
+      {ratingsEl}
+      {reviewEl}
     </div>
   );
 };
 
-SimpleSummary.displayName = 'SimpleSummary';
+ProductSummary.displayName = 'ProductSummary';
 
-SimpleSummary.propTypes = {
+ProductSummary.propTypes = {
+  price: PropTypes.string,
   title: PropTypes.string.isRequired,
-  url: PropTypes.string.isRequired,
+  reviewCount: PropTypes.number,
+  rating: PropTypes.number,
+  maxRating: PropTypes.number,
+  minRating: PropTypes.number,
 };
 
-export default SimpleSummary;
+export default ProductSummary;
